+++ source/extensions/omni.isaac.orbit/docs/CHANGELOG.rst
@@ -1,6 +1,27 @@
 Changelog
 ---------
 
+0.10.12 (2024-01-10)
+~~~~~~~~~~~~~~~~~~~~
+
+Fixed
+^^^^^
+
+* Fixed indexing of source and target frames in the :class:`omni.isaac.orbit.sensors.FrameTransformer` class.
+  Earlier, it always assumed that the source frame body is at index 0. Now, it uses the body index of the
+  source frame to compute the transformation.
+
+Deprecated
+^^^^^^^^^^
+
+* Renamed quantities in the :class:`omni.isaac.orbit.sensors.FrameTransformerData` class to be more
+  consistent with the terminology used in the asset classes. The following quantities are deprecated:
+
+  * ``target_rot_w`` -> ``target_quat_w``
+  * ``source_rot_w`` -> ``source_quat_w``
+  * ``target_rot_source`` -> ``target_quat_source``
+
+
 0.10.11 (2024-01-08)
 ~~~~~~~~~~~~~~~~~~~~